requirements.txt
toml
requests==2.31.0
openai==0.27.8
newsapi-python
markdown
pandas
boto3==1.34.34
